fix(tests): Moved test utilities to helper modules
Create modules for creating games, users, players for reuse in tests
