+++ test/models/concerns/player_stats/session_stats_test.rb
@@ -1,9 +1,11 @@
 require 'test_helper'
 
 class SessionStatsTest < ActionController::TestCase
+  include GameCreation, PlayerCreation, SessionCreation
+
   test 'should return number of sessions played' do
     player = create_player
-    create_session(create_game, player)
+    create_session(create_game, winning_players: [player])
 
     session_count = PlayerStatistic.new(player.id).session_count
 
@@ -12,34 +14,10 @@ class SessionStatsTest < ActionController::TestCase
 
   test 'should count sessions with duplicate player once' do
     player = create_player
-    create_session(create_game, player, player)
+    create_session(create_game, winning_players: [player, player])
 
     session_count = PlayerStatistic.new(player.id).session_count
 
     assert_equal 1, session_count
   end
-
-  private
-    def create_player(name = 'test')
-      player = Player.new(:name => name)
-      player.save
-      return player
-    end
-
-    def create_game
-      game = Game.new(:name => 'test', :game_type => Game.game_types[:mixed])
-      game.save
-      return game
-    end
-
-    def create_session(game, *players)
-      session = Session.new(:game_id => game.id)
-      session.save
-
-      players.each do |player|
-        session_player2 = SessionPlayer.new(:session_id => session.id, :player_id => player.id)
-        session_player2.placing = 2
-        session_player2.save
-      end
-    end
 end
