+++ test/support/player_creation.rb
@@ -1,0 +1,7 @@
+module PlayerCreation
+  def create_player(name = 'test')
+    player = Player.new(:name => name)
+    player.save
+    return player
+  end
+end
